+++ src/main/groovy/com/ullink/NuGetSpec.groovy
@@ -8,6 +8,7 @@ import groovy.util.slurpersupport.GPathResult
 import groovy.xml.XmlUtil
 import org.gradle.api.tasks.Exec
 import org.gradle.api.tasks.Input
+import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.OutputFile
 
 class NuGetSpec extends Exec {
@@ -20,6 +21,7 @@ class NuGetSpec extends Exec {
         generateNuspecFile()
     }
 
+    @Internal
     File getTempNuspecFile() {
         new File(temporaryDir, project.name + '.nuspec')
     }